[Feat/#237] Refactor Package Structure
