+++ src/main/java/com/example/waggle/domain/board/presentation/dto/story/StoryRequest.java
@@ -16,5 +16,7 @@ public class StoryRequest {
     @Size(max = 500)
     private String content;
     private List<String> hashtagList;
+
+    @Size(min = 1, message = "미디어는 최소 1개 이상 필요합니다.")
     private List<String> mediaList;
 }
