+++ src/main/java/com/example/waggle/global/config/MultipartJackson2HttpMessageConverter.java
@@ -1,4 +1,4 @@
-package com.example.waggle.global.config.spring;
+package com.example.waggle.global.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.lang.reflect.Type;
