+++ src/main/java/com/example/waggle/global/config/SchedulerConfig.java
@@ -1,4 +1,4 @@
-package com.example.waggle.global.config.spring;
+package com.example.waggle.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.SchedulingConfigurer;
